pyproject.toml
[project]
name = "phoenix-lit"
version = "0.1.0"
description = ""
authors = [
    {name = "mahaElmorsy1",email = "mom@dgtera.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "typer (>=0.20.0,<0.21.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "arabic-reshaper (>=3.0.0,<4.0.0)",
    "python-bidi (>=0.6.7,<0.7.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "opencv-python (==4.8.0.74)",
    "numpy (>=1.26,<2)",
    "mediapipe (>=0.10.21,<0.11.0)",
    "pyttsx3 (>=2.90,<2.100)",
    "scikit-learn (>=1.4,<1.6)",
    "joblib (>=1.3,<1.5)",
    "matplotlib (>=3.10.8,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
phoenixlit = "phoenixlit.cli:app"

[tool.poetry]
packages = [{ include = "phoenixlit", from = "src" }]